requirements.txt
Django==5.2.7
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.3
drf-yasg==1.21.7
django-extensions==3.2.3
psycopg2-binary==2.9.9
redis==5.0.2
django-redis==5.4.0
whitenoise==6.6.0
gunicorn==21.2.0
python-dotenv==1.0.1
Pillow==10.3.0
django-storages==1.14.2
boto3==1.34.25
celery==5.3.6
django-celery-beat==2.6.0

# Development
debugpy==1.8.1
ipython==8.22.2